+++ views/registroProducto.html
@@ -130,7 +130,7 @@
     <!-- MDB -->
     <!-- <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/mdb-ui-kit/3.3.0/mdb.min.js"></script> -->
     <!-- Await + Modal + Efecto Spinner  -->
-    <script type="module" src="/js/configDB.js"></script>
+ 
     <script type="module" src="/js/registroProducto.js"></script>
 </body>
